+++ src/main/java/net/thesquire/backroomsmod/BackroomsModClient.java
@@ -37,7 +37,8 @@ public class BackroomsModClient implements ClientModInitializer {
                 ModBlocks.PIPE_BLOCK,
                 ModBlocks.ELEVATOR_BUTTON,
                 ModBlocks.ELEVATOR_DOOR,
-                ModBlocks.OFFICE_WINDOW);
+                ModBlocks.OFFICE_WINDOW,
+                ModBlocks.FRAMED_WINDOW);
 
         BackroomsMod.LOGGER.info("Client initialization finished for " + BackroomsMod.MOD_ID);
     }