+++ src/main/java/net/thesquire/backroomsmod/world/ModBiomes.java
@@ -122,6 +122,8 @@ public class ModBiomes {
                 placedFeatureRegistryEntryLookup.getOrThrow(ModPlacedFeatures.LEVEL_4_THIN_CROOKED_WALL_PLACED_KEY));
         level4BiomeBuilder.feature(GenerationStep.Feature.LOCAL_MODIFICATIONS,
                 placedFeatureRegistryEntryLookup.getOrThrow(ModPlacedFeatures.LEVEL_4_FLUORESCENT_LIGHT_PLACED_KEY));
+        level4BiomeBuilder.feature(GenerationStep.Feature.LOCAL_MODIFICATIONS,
+                placedFeatureRegistryEntryLookup.getOrThrow(ModPlacedFeatures.LEVEL_4_WINDOWS_PLACED_KEY));
 
         // level 11 biome settings
         GenerationSettings.Builder level11BiomeBuilder = new GenerationSettings.Builder();