+++ src/main/java/net/thesquire/backroomsmod/world/feature/ModFeatures.java
@@ -20,6 +20,7 @@ public class ModFeatures {
     public static final RegistryKey<Feature<?>> FLAT_ORE_FEATURE_KEY = registerKey("flat_ore_feature");
     public static final RegistryKey<Feature<?>> LOOT_CHEST_KEY = registerKey("loot_chest");
     public static final RegistryKey<Feature<?>> PIPE_NETWORK_KEY = registerKey("pipe_network");
+    public static final RegistryKey<Feature<?>> WINDOW_KEY = registerKey("window");
 
     public static Feature<ModSimpleWallFeatureConfig> WALL;
     public static Feature<ModThinWallFeatureConfig> THIN_WALL;
@@ -29,6 +30,7 @@ public class ModFeatures {
     public static Feature<OreFeatureConfig> FLAT_ORE_FEATURE;
     public static Feature<ModLootChestFeatureConfig> LOOT_CHEST;
     public static Feature<ModPipeNetworkFeatureConfig> PIPE_NETWORK;
+    public static Feature<ModWindowFeatureConfig> WINDOW;
 
     public static void registerModFeatures() {
         BackroomsMod.LOGGER.info("Registering mod features for " + BackroomsMod.MOD_ID);
@@ -41,6 +43,7 @@ public class ModFeatures {
         FLAT_ORE_FEATURE = register(FLAT_ORE_FEATURE_KEY.getValue().getPath(), new ModFlatOreFeature(OreFeatureConfig.CODEC));
         LOOT_CHEST = register(LOOT_CHEST_KEY.getValue().getPath(), new ModLootChestFeature(ModLootChestFeatureConfig.CODEC));
         PIPE_NETWORK = register(PIPE_NETWORK_KEY.getValue().getPath(), new ModPipeNetworkFeature(ModPipeNetworkFeatureConfig.CODEC));
+        WINDOW = register(WINDOW_KEY.getValue().getPath(), new ModWindowFeature(ModWindowFeatureConfig.CODEC));
     }
 
     public static RegistryKey<Feature<?>> registerKey(String name) {